I'm sorry, should've been split into several commits:
* Use u32 for width instead of u16
* In order to fix bug where registers > 65535 bits weren't being allocated at all
* stub out currentvl_read
* add tests for slice mask and v_set

+++ brig/kernel/src/host/dbt/emitter.rs
@@ -114,15 +114,15 @@ pub trait Emitter<A: Alloc> {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Type {
-    Unsigned(u16),
-    Signed(u16),
-    Floating(u16),
+    Unsigned(u32),
+    Signed(u32),
+    Floating(u32),
     Bits,
     Tuple,
 }
 
 impl Type {
-    pub fn width(&self) -> u16 {
+    pub fn width(&self) -> u32 {
         match self {
             Type::Unsigned(w) | Type::Signed(w) | Type::Floating(w) => *w,
             Type::Bits => 64, // todo: should this be the runtime length?
